requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
idna==3.11
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.3
packaging==26.1
pandas==3.0.1
pillow==12.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
seaborn==0.13.2
six==1.17.0
tzdata==2025.3
urllib3==2.6.3
